requirements.txt
Django==4.0.2
django-crispy-forms==1.14.0
Pillow
psycopg2-binary
six==1.16.0
whitenoise==6.0.0
django-environ==0.9.0
djangorestframework==3.13.1
django-filter==22.1
drf-spectacular==0.23.1
pre-commit==2.20.0
django-rest-passwordreset==1.3.0
boto3==1.26.66
django-storages==1.13.2
django-cors-headers==4.3.1